+++ eras/byron/ledger/impl/src/Cardano/Chain/Update/Validation/Interface.hs
@@ -4,6 +4,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeApplications #-}
 
 -- | Blockchain interface validation rules.
 module Cardano.Chain.Update.Validation.Interface (
@@ -211,7 +212,7 @@ instance DecCBOR Error where
       1 -> checkSize 2 >> Voting <$> decCBOR
       2 -> checkSize 2 >> Endorsement <$> decCBOR
       3 -> checkSize 2 >> NumberOfGenesisKeysTooLarge <$> decCBOR
-      _ -> cborError $ DecoderErrorUnknownTag "Interface.Error" tag
+      _ -> cborError $ DecoderErrorUnknownTag "Interface.Error" $ fromIntegral @Word8 @Word tag
 
 -- | Signal combining signals from various rules
 data Signal = Signal